% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{api_uri}
\alias{api_uri}
\title{Retuirve the base URI for the homr API}
\usage{
api_uri(..., root = "http://www.ncdc.noaa.gov/homr/services/station")
}
\arguments{
\item{...}{path segments passed to \code{file.path}}

\item{root}{char, the root address}
}
\value{
char URI
}
\description{
Retuirve the base URI for the homr API
}
