% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homr-package.R
\docType{package}
\name{homr-package}
\alias{homr-package}
\alias{_PACKAGE}
\alias{homr}
\title{homr package}
\description{
Querying and accessing NOAA's Historical Observing Metadata Repository.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/homr}
}

}
