% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_search}
\alias{query_search}
\title{Query homr using the search endpoint}
\usage{
query_search(..., include_definitions = FALSE, base_uri = api_uri("search"))
}
\arguments{
\item{...}{elements of the query to be cast and \code{character} type}

\item{include_definitions}{logical, if TRUE add definitions as an attribute
of the returned value}

\item{base_uri}{char, the base uri}
}
\value{
tibble
}
\description{
Query homr using the search endpoint
}
